server/requirements.txt
openai
python-dotenv
langchain
langchain-community
langchain-openai
langchain-text-splitters
bs4
requests
beautifulsoup4
playwright
unstructured
selenium
chromadb
langchain-chroma
fastapi
uvicorn
pydantic